NTFCollection refactor and minor fixes
