+++ src/Components/Modals/LoginRegister.tsx
@@ -315,6 +315,7 @@ const LoginRegister = ({
                 aria-labelledby="modal-modal-title"
                 aria-describedby="modal-modal-description"
                 disableEscapeKeyDown={persistent}
+                disableAutoFocus
                 slotProps={persistent ? {backdrop:{sx:{background: 'rgba(0, 0, 0)'}}} : {}}
             >
                 <Box sx={style}>
